requirements.txt
pandas>=2.0.0
numpy>=1.24.0
cohere>=5.2.0
google-generativeai>=0.5.0
langchain>=0.1.0
langchain-cohere>=0.1.0
python-dotenv>=1.0.0
rich>=13.0.0
# streamlit
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0